+++ packages/ui/src/components/organisms/Form/DynamicForm/fields/EntityFieldGroup/providers/EntityFieldProvider/types.ts
@@ -1,0 +1,7 @@
+import { TEntityFieldGroupType } from '../../EntityFieldGroup';
+
+export interface IEntityFieldProviderContext {
+  entityFieldGroupType?: TEntityFieldGroupType;
+  entityId?: string;
+  isSyncing: boolean;
+}